requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# HTTP Requests & Git Operations
requests==2.31.0
gitpython==3.1.37

# Code Analysis Tools
bandit==1.7.5
pylint==2.17.5

# Environment & Configuration
python-dotenv==1.0.0

# Database (SQLite - built-in Python)
# No additional dependencies needed

# AI/LLM Integration (Groq API)
# Uses requests library - no additional dependencies

# Optional: Testing (if needed)
# pytest==7.4.3
